refactor: minor cleanups and correctness fixes
Grant Read permission for SortMergeJoin in the security identity check.
Replace superfluous u32 cast in dict-encode cardinality check. Swap
heap-allocated test buffers for stack arrays in msgpack_scan field and
index tests. Simplify nested fixarray construction in reader depth test.
Fix unnecessary borrow in sidecar verify_entry call and update a
floating-point test value to an exactly representable f64.

+++ nodedb-columnar/src/memtable.rs
@@ -558,7 +558,7 @@ impl ColumnarMemtable {
     /// columns. Improves compression and enables integer-based GROUP BY /
     /// WHERE evaluation on the resulting segment.
     pub fn drain_optimized(&mut self) -> (ColumnarSchema, Vec<ColumnData>, usize) {
-        self.try_dict_encode_columns(DICT_ENCODE_MAX_CARDINALITY as u32);
+        self.try_dict_encode_columns(DICT_ENCODE_MAX_CARDINALITY);
         self.drain()
     }
 }
